requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cssselect==1.1.0
cssutils==1.0.2
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-filter==2.3.0
django-regex==0.4.1
django-rest-auth==0.9.5
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
emails==0.6
gunicorn==20.0.4
idna==2.10
lxml==4.5.2
oauthlib==3.1.0
Pillow==7.2.0
premailer==3.7.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python-imap==1.0.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
